requirements.txt
streamlit>=1.28.0
gspread>=5.12.0
google-auth>=2.23.0
pandas>=2.0.0
plotly>=5.18.0
pydantic>=2.0
pytest>=7.0
